extract helper tests
